src/requirement.txt
cassandra-driver==3.0.0
mmh3==2.2
bitarray==0.8.1
